Fix Visitor Profile link disappears on following pages in visitor log
fix #14998

+++ plugins/Live/Visualizations/VisitorLog.php
@@ -108,8 +108,8 @@ class VisitorLog extends Visualization
             $this->config->custom_parameters['totalRows'] = 10000000;
         }
 
-        $this->config->custom_parameters['smallWidth'] = (1 == Common::getRequestVar('small', 0, 'int'));
-        $this->config->custom_parameters['hideProfileLink'] = (1 == Common::getRequestVar('hideProfileLink', 0, 'int'));
+        $this->config->custom_parameters['smallWidth'] = (int)(1 == Common::getRequestVar('small', 0, 'int'));
+        $this->config->custom_parameters['hideProfileLink'] = (int)(1 == Common::getRequestVar('hideProfileLink', 0, 'int'));
         $this->config->custom_parameters['pageUrlNotDefined'] = Piwik::translate('General_NotDefined', Piwik::translate('Actions_ColumnPageURL'));
 
         if (Common::getRequestVar('inPopover', '0') === '0') {
